# docs/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHCW2026 OpenClaw API V2",
    "version": "2.2.0",
    "description": "Machine-readable OpenAPI spec for SHCW2026 /api/v2 endpoints. V2 preserves V1 behavior and adds response header X-API-Version plus standardized error code field, event registration audience endpoints, and expanded user management operations."
  },
  "servers": [
    {
      "url": "https://shcw2026.vercel.app",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "xApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v2/users/{id}": {
      "parameters": [
        { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "tags": ["Users"],
        "summary": "Get user detail",
        "responses": {
          "200": { "description": "User detail", "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "$ref": "#/components/responses/ErrorUnauthorized" },
          "403": { "$ref": "#/components/responses/ErrorForbidden" },
          "404": { "$ref": "#/components/responses/ErrorNotFound" },
          "500": { "$ref": "#/components/responses/ErrorInternal" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "role": { "type": "string" },
                  "status": { "type": "string" },
                  "phone": { "type": "string" },
                  "title": { "type": "string" },
                  "bio": { "type": "string" },
                  "avatar": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated", "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "$ref": "#/components/responses/ErrorBadRequest" },
          "401": { "$ref": "#/components/responses/ErrorUnauthorized" },
          "403": { "$ref": "#/components/responses/ErrorForbidden" },
          "404": { "$ref": "#/components/responses/ErrorNotFound" },
          "409": { "$ref": "#/components/responses/ErrorConflict" },
          "500": { "$ref": "#/components/responses/ErrorInternal" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authorization: Bearer sk_oc_..."
      },
      "xApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "x-api-key: sk_oc_..."
      }
    },
    "headers": {
      "XApiVersion": {
        "description": "API version marker",
        "schema": {
          "type": "string",
          "enum": ["v2"]
        }
      }
    },
    "responses": {
      "ErrorBadRequest": {
        "description": "Bad Request",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Invalid request", "code": "BAD_REQUEST" } } }
      },
      "ErrorUnauthorized": {
        "description": "Unauthorized",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Unauthorized", "code": "UNAUTHORIZED" } } }
      },
      "ErrorForbidden": {
        "description": "Forbidden",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Forbidden", "code": "FORBIDDEN" } } }
      },
      "ErrorNotFound": {
        "description": "Not Found",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Not found", "code": "NOT_FOUND" } } }
      },
      "ErrorConflict": {
        "description": "Conflict",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Conflict", "code": "CONFLICT" } } }
      },
      "ErrorInternal": {
        "description": "Internal Server Error",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Internal server error", "code": "INTERNAL_ERROR" } } }
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "message": { "type": "string" },
          "data": {}
        },
        "additionalProperties": true
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "code"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "code": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  }
}